samples/python/voice-live-agent/pyproject.toml
[project]
name = "voice-live-agent"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
  "fastapi[standard]",
  "azure-communication-callautomation",
  "azure-identity",
  "websockets",
  "python-dotenv",
]
